+++ src/components/Header/Header.module.scss
@@ -1,4 +1,4 @@
-@import 'variables';
+@import '../../styles/variables';
 
 .poop {
   font-size: 2rem;
