+++ src/components/Header/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as styles from '../styles/Header.module.scss';
+import * as styles from './Header.module.scss';
 
 export default function Header() {
   return (
