+++ src/components/Layout/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from './Header';
-import Footer from './Footer';
+import Header from '../Header';
+import Footer from '../Footer';
 
 export default function Layout({ children }) {
   return (
